Update tokenization.py
The BERT repo uses Tensorflow 1 and thus a few of the functions have been moved/changed/renamed in Tensorflow 2. In order for the BERT tokenizer to be used, one of the lines in the repo that was just cloned needs to be modified to comply with Tensorflow 2. Line 125 in the BERT tokenization.py file must be changed as follows:

From => with tf.gfile.GFile(vocab_file, "r") as reader:

To => with tf.io.gfile.GFile(vocab_file, "r") as reader: